moved arrow to a new component
