+++ src/atomic/atoms/Arrow.tsx
@@ -1,0 +1,38 @@
+"use client";
+import { FC } from "react";
+import styled from "styled-components";
+
+export interface ArrowProps {
+    startingRow: number;
+    endingRow: number;
+    startingColumn: number;
+    endingColumn: number;
+}
+
+interface IArrow {
+    $startingRow: number;
+    $endingRow: number;
+    $startingColumn: number;
+    $endingColumn: number;
+}
+
+const StyledTestArrow = styled.div<IArrow>`
+    grid-row-start: ${(props) => props.$startingRow};
+    grid-row-end: ${(props) => props.$endingRow + 1};
+    grid-column-start: ${(props) => props.$startingColumn};
+    grid-column-end: ${(props) => props.$endingColumn + 1};
+    background-color: yellow;
+    width: 100%;
+    height: 50%;
+`;
+
+export const Arrow: FC<ArrowProps> = ({ startingRow, endingRow, startingColumn, endingColumn }) => {
+    return (
+        <StyledTestArrow
+            $startingRow={startingRow}
+            $endingRow={endingRow}
+            $startingColumn={startingColumn}
+            $endingColumn={endingColumn}
+        />
+    );
+};
