+++ src/atomic/organisms/TalentTree.tsx
@@ -4,6 +4,7 @@ import styled from "styled-components";
 import { SpecTalent } from "@/data/classData";
 import { TalentNode } from "../atoms/TalentNode";
 import Image from "next/image";
+import { Arrow } from "../atoms/Arrow";
 
 interface IStyledContainer {
     $backgroundImage: string;
@@ -48,42 +49,6 @@ interface TalentTreeProps {
     specData: SpecTalent[];
 }
 
-const StyledVerticalTestArrow = styled.div`
-    grid-row-start: 4;
-    grid-row-end: 5;
-    grid-column-start: 3;
-    grid-column-end: 3;
-    background-color: yellow;
-    width: 10px;
-    height: 100%;
-`;
-
-const StyledHorizontalTestArrow = styled.div`
-    grid-row-start: 5;
-    grid-row-end: 5;
-    grid-column-start: 4;
-    grid-column-end: 5;
-    background-color: yellow;
-    width: 100%;
-    height: 10px;
-`;
-
-interface IArrow {
-    $startingRow: number;
-    $endingRow: number;
-    $startingColumn: number;
-    $endingColumn: number;
-}
-const StyledTestArrow = styled.div<IArrow>`
-    grid-row-start: ${(props) => props.$startingRow};
-    grid-row-end: ${(props) => props.$endingRow + 1};
-    grid-column-start: ${(props) => props.$startingColumn};
-    grid-column-end: ${(props) => props.$endingColumn + 1};
-    background-color: yellow;
-    width: 100%;
-    height: 50%;
-`;
-
 export const TalentTree: FC<TalentTreeProps> = ({ specName, specData, specIcon, specBackground }) => {
     if (specName === "Beast Mastery") {
         console.log("specBackground:", specBackground);
@@ -110,18 +75,16 @@ export const TalentTree: FC<TalentTreeProps> = ({ specName, specData, specIcon, 
                                 key={node.talentId}
                             />
                             {node.unlocks && node.unlocks?.length > 0 ? (
-                                <StyledTestArrow
-                                    $startingRow={node.talentRow}
-                                    $endingRow={node.unlocks[0]}
-                                    $startingColumn={node.talentcolumn}
-                                    $endingColumn={node.unlocks[1]}
+                                <Arrow
+                                    startingRow={node.talentRow}
+                                    endingRow={node.unlocks[0]}
+                                    startingColumn={node.talentcolumn}
+                                    endingColumn={node.unlocks[1]}
                                 />
                             ) : null}
                         </>
                     );
                 })}
-
-                {/* <StyledHorizontalTestArrow> </StyledHorizontalTestArrow> */}
             </TalentGrid>
         </MainContainer>
     );
